Re-use a pre-composed email template for cloud share
