+++ public/js/cloud_share_modal.js
@@ -84,7 +84,7 @@ export default class CloudShareModal extends React.Component {
   renderResults() {
     const { shareableurl } = this.state;
 
-    return <ShareURLComponent url={shareableurl} />;
+    return <ShareURLComponent url={shareableurl} querydb={this.props.querydb} program={this.props.program} queryLength={this.props.queryLength} />;
   }
 
   renderError() {
