+++ public/js/share_url.js
@@ -1,6 +1,7 @@
 import React, { useState } from 'react';
+import asMailtoHref from './mailto';
 
-const ShareURLComponent = ({ url }) => {
+const ShareURLComponent = ({ querydb, program, queryLength, url }) => {
   const [copied, setCopied] = useState(false);
 
   const copyToClipboard = () => {
@@ -13,7 +14,7 @@ const ShareURLComponent = ({ url }) => {
       <input name="shareableUrl" type="text" value={url} readOnly />
       <div className="actions">
         <button className="btn btn-primary" onClick={copyToClipboard}>{copied ? 'Copied!' : 'Copy to Clipboard'}</button>
-        <a href={`mailto:?subject=SequenceServer result&body=${encodeURIComponent(url)}`}>Share via email</a>
+        <a href={asMailtoHref(querydb, program, queryLength, url, true)}>Share via email</a>
       </div>
     </div>
   );
